🐛 [Fix] Custom built version of gpt library
- To support settings per function
- Fix password reset token to check existing user email
- [WIP] gpt's notice for each baumann room

+++ java/src/main/java/app/beautyminder/service/auth/UserService.java
@@ -191,7 +191,7 @@ public class UserService {
 
         // Delete user profile picture from S3
         String userProfileImage = user.getProfileImage();
-        if (!userProfileImage.isEmpty() &&
+        if (userProfileImage != null && !userProfileImage.isEmpty() &&
                 !List.of(defaultUserProfilePic, defaultAdminProfilePic).contains(userProfileImage)) {
             fileStorageService.deleteFile(userProfileImage);
         }
